requirements.txt
beautifulsoup4
discord
python-dotenv
pymupdf
requests